# Optimization: fuse linesearch_zero_jv and linesarch_init_jv