For humanoid, but I'm seeing good stuff for many other things as well. This is a tiny piece of an effort to work on all of these tiny kernels, so sometimes it might not even show up in the overall runtime. But looking at low-level traces, there is a big difference.

Time spent in the 2 kernels before/after for 8k humanoids: 11ms -> 1.5ms.

main:

```
Summary for 8192 parallel rollouts

 Total JIT time: 1.18 s
 Total simulation time: 0.16 s
 Total steps per second: 2,555,648
 Total realtime factor: 12,778.24 x
 Total time per step: 391.29 ns

```
this:

```
Summary for 8192 parallel rollouts

 Total JIT time: 1.12 s
 Total simulation time: 0.15 s
 Total steps per second: 2,773,893
 Total realtime factor: 13,869.46 x
 Total time per step: 360.50 ns

```